Cleaning up DayMetric initialization.
Yesterday will now reference the object from the previous day.

In addition, I cleaned up the today initialization by first querying for yesterday and if it exists then
it will set the pull count.Then we will query for today, if it exists.If it does, it will overwrite all previous
day metrics and if does not exist then it will use whatever current default values were set and update the date value accordingly.