Need to use dotenv-cli package instead of dotenv
While folks were QAing the PR, it seems that some of the package json
scripts failed because npx could not find the `dotenv` cli command from
the dotenv package. This is as expected because the dotenv package does
not have a cli command; however, the dotenv-cli package does have this
command and therefore should be used.

It worked locally for me because I had globally installed the package.