feat(docker-compose): created docker-compose with mysql db to start
Created a docker-compose file with a mysql(db) service that only has a
root password defined and no databases defined, because the prisma
client will generate the dev and test databases but can only do so with
the root user since non-defined user would need to be granted access to
both  databases. This would be messy to do without added an additional
script to run a SQL query and update permissions, but then you would
need to keep track of the users created for the env file that prisma
would use.