feat(docker-compose): added main app service
Added the service 'app' which uses the Dockerfile perviously created.
It also will use the '--env=file' option to pass the
'$DOCKER_DATABASE_URL' variable to the container which the prisma client
in the container will use to connect the MySQL DB running on the other
container. This will allow the use of one env file for each staging
without needed to duplicate it for local or docker containers.

I also updated the env template to reflect the changes