fix(jest): tests were not exiting properly
There was a warning about jest hanging because prisma was not
disconnecting as expected when the tests were done.
Therefore, I added teardown script that jest runs after all
the tests have finished.

Also lint-staged complained about module not defined in jest.config.js
so I renamed it to typescript with `git mv`.