refactor(docker-compose): changed network mode to host
I changed the network mode to host to let the app container share
the host’s networking stack. This approach means localhost inside a
container resolves to the physical host, instead of the container
itself. The reason why this approach was taken was primarily due to
reducing the number of `.env` files that would be needed by the
application. In which,
we only need to specify the DATABASE_URL once instead of declaring
different DATABASE_URLs based on whether the applicaiton was running in
the docker network or locally--the approach previously taken.

In addition, I removed the environment variable for the app container
because it could not be overridden when then trying to run
`prisma:test_setup`. This meant that the tests would be using the dev db
and therefore delete the data in those tables after completing the tests

Also, now that the application is running on the local network, and we
can connect to the database without a separate DATABASE_URL for docker,
we no longer need the prisma studio service.