feat(parse): add granular pull metrics
Adding a column for number of dev blocks on a pull request as well as
number of qa stamps. This will be used to quickly query pull requests
for those values instead of running multiple conditionals to calculate
it later.

This also changes the name for interaction counts and qa ready counts to
'agg' as it will be aggregation of those events later on.