feat(parse): add pull's head commit index table
This adds a table to keep track of the head commit associated with
a given pull request.

In order to keep track of the pull's current state we also need to
refrence the head commit but because
a pull request can have many commits, there are now two points from the
pull request table to the commits table.
Therefore, separating out this relationship into its own table
was easier to do, especially with Prisma.