feat(parse): add gh graphql queries
In order to populate the new tables and columns, we now need additional
queries to the GH GraphQL.

This adds those querires as well as additional tests to validate them.

That said, the tests failed with the Commit Fragment part because the
different types, PullRequestCommit and PullRequestReview
return conflicting types for Commit. Where PullRequestCommit returns
a possible nullish Commit object