refactor(comment): clean up dev block parsing
Since each comment is treated as an individual event, then we no longer need to
iterate through an array of comments.

In similar fashion as the QAed parsing, we no longer need to check for
a null comment or bodyText.

Aside from that, while also checking the comment for a dev_block match
we can check for an un_dev_block match. If there are no matches we will
return null.