feat(ph_controller): record qa ready if qa required
If the Pull Request has no QA required, then the commits can never
be 'QA Ready'. Therefore, lets check if a commit is QA ready by aslo
validating that QA is required for the Pull Request.

In addition, recording the state change to 'Non QA Ready'
does not serve a purpose if the Pull Request has no QA Required either.