feat(backill): backfill PullRequest
Backfill the PullRequest with the count of each event type recorded.

Also set the PullRequest state based off the head commit state.
The last dev block state is tied with a particular commit so we cannot
set all consecutive commits with a 'dev_blocked' state since the target of the
dev block comment was for a particular commit.