feat(index): add timeline func with sanitization
Now that the timeline parsing logic has been hammered out.
We will integrate it into the index file.

That said, we will need to query github for open pulls with the
timeline fields as part of the request, so we need to change the query to be
used on each repo.

To help filter out the nested possibly null types from the github nodes
I created a helper function to filter out the Maybe<Maybe<T>[]> values with
a predicate type.