feat(graphql): add head commit id field
With the additional head_commit_id column, we now need to request
the head commit id from GitHub.

However, it is possible that the head_commit_id maybe null if the
pull request is already merged. Therefore, let's make sure to also set
the head commit id when backfilling the pull request.