refactor(date): use date-fns instead
Moment.js is considered completed by the devs.
After taking a look at some alternatives,
it seems date-fns will be a good alternative as it supports typescript
and is used in a lot of public repositories --
https://vhudyma-blog.eu/the-best-momentjs-alternatives/ --
date-fns seems to be a good alternative.

Therefore, replacing current usage of moment and since we no longer need
to get the current date for DayMetrics, we can
now remove date-and-time as well.