fix(save): handle concurrent saving prisma transactions
When trying to save to the database, a foreign key constraint error was
thrown because the commits were still not saved to the database when
the events were getting saved.

Instead of trying to create a new promise array for the other db
transactions. We can use prisma's transaction api to manage some of
the batching. This requires the setting
to be defined in the Prisma Schema file.

In addition, instead of using map, I choose to use a for loop to respect
the awaits and therefore not need the Promise array.