feat(db_pul): add pull state functionality
It was getting tedious setting and getting the pull state for control
statements. Which was also an issue because we needed to also know about
a different pull state before logging pull events. For instance, we can
have another entry for `non_qa_ready` if the commit was previously QAed
or Dev Blocked and a new QAed or Dev Blocked entry was created.

I therefore, pulled the local pull state variables used in
pull_histroy_controller into db_pull to be a part of the Pull model.
In order to make it easier to work with, I also created
setters and getters for those properties.

As a result of this, I neded to refactor a lot of the functions to make
use or include a pull argument to retrieve the necessary states.
This reduced a lot of the reseting of the pull state variables.