refactor(pull_history_controller): reorder lines and conditional
Changing the conditional check to be QAReady state instead of the
current dev block state to keep it consistent with the logic that
the Dev Block function is using as well in
CheckAndRecordDevBlockSignature.

Then to make it easier to ready, I flipped the conditional checks for
the Un Dev Block logic. I also moved the setting of the pull states to
the end of the blocks to make it easier to read.