refactor(comment): parse the comments in ctx with signature
When creating the process flow chart with @jordycosta, he suggested
parsing the comments within the checking of the signatures and since we
abstracted the parsing for specific signatures into
individual functions, there was no reason to not directly use them.

Therefore, I removed the parseComment function that intergrates
checking all signatures and made individual calls to each signature.