feat: day metrics test
In order to properly test day metrics, we need to mock the response
from Prisma which requires jest-mock-extended to be installed.
In addition, we need to create a singleton class for prisma for the
tests and the declaration of it within the Jest config setup.

As a result of the prismamock we are also having a cyclic warning that
Prisma docs also talk about and can be resolved by removing strict null
checks in the TS config.